Replaced position classes with an enum
