--- src/main/java/unq/dapp/supergol/model/Forward.java
@@ -1,8 +1,0 @@
-package unq.dapp.supergol.model;
-
-public class Forward implements Position {
-  @Override
-  public int scoreFor(Player player, Match match) {
-    return match.goalsOf(player);
-  }
-}
